Last Commit included:
Changed MainActivity-> front page, 3 buttons to Test Bed, a Single Table Q List and Multi Table Q list

Previous MainActivity is now QuestionListActivity, used for both lists.

Changed Question Server ( to have 2 lists, and needs to be updated to switch for the right list to load

TestBedActivity mimicked QuestionActivity, but uses all tables and i added a TestTableAdapter to display a nice list of all the tables

TestAnswerActivity mimicked AnswerCheckerActivity, but instead only displays the query result and a Retry or Return to Front buttons.